requirements.txt
Django>=1.8
djangorestframework>=3.0
python-dateutil
python-magic
six
wheel
